+++ lib/src/extensions/enum_extensions.dart
@@ -3,8 +3,11 @@ import 'string_extensions.dart';
 extension EnumExtn on Enum {
   /// Returns enum value as title case with space
   ///
-  /// Example: veryLongWord => Very Long Word.
-  ///
+  /// Example:
+  /// ```dart
+  /// enum Words { veryLongWord }
+  /// print(Words.veryLongWord.titleCase); // prints "Very Long Word"
+  /// ```
   /// See string extension function: [titleCase]
   String get titleCase => name.splitCamelCaseWord().titleCase;